Hotfix for TextHandler class to be able to reliably debug the Level and Walldoor classes
Created a different way to get each Hanja from TextHandler class. Previously there was no variable storing the length of each Sino-korean word.

You might've been able to create a seperate list for this, but then you would also have to add to it in parallel and associate it with the right index in the hanjaArray list. The hanjaArray along with its hanjaIndex list have now been replaced by an Object ArrayList called hanjaContainer.

This list stores Objects that are instances of Boolean and Character, where each Hanja is stored as a char and each non-Hanja is stored as a boolean. The Level class now checks each syllable in the hanjaContainer list for any instance of Character. If there is than we can tell that it's Hanja.