Added TSV for definitions
PWindow now uses a Table(Excel spreadsheet) loaded with a TSV-file (Tab Seperated Values).
PWindowText object instances are now used in PWindow to be able to display definitions later. The currently collided Walldoor, or if no collision detected the collidedOnce, of highest index should later be displayed in PWindow. Still a WIP